Modify Matlab's fmincon optimization function call to converge better
Added more constraints to the optimization function so that Matlab can estimate the gradient of each constraint separately (before, the constraint was that minimum of all drone distances were higher than threshRisk, so it was hard for the algorithm to understand how each parameter played a role in satisfying that constraint).
Now, each drone-drone and drone-wall constraint is a separate constraint, and Matlab finds the solution much quicker (and better! :D)